+++ lang/attribute/program/src/declare_program/common.rs
@@ -1,6 +1,7 @@
 use anchor_lang_idl::types::{
-    Idl, IdlArrayLen, IdlDefinedFields, IdlField, IdlGenericArg, IdlRepr, IdlSerialization,
-    IdlType, IdlTypeDef, IdlTypeDefGeneric, IdlTypeDefTy,
+    Idl, IdlArrayLen, IdlDefinedFields, IdlField, IdlGenericArg, IdlInstructionAccountItem,
+    IdlInstructionAccounts, IdlRepr, IdlSerialization, IdlType, IdlTypeDef, IdlTypeDefGeneric,
+    IdlTypeDefTy,
 };
 use proc_macro2::Literal;
 use quote::{format_ident, quote};
@@ -39,11 +40,11 @@ pub fn gen_accounts_common(idl: &Idl, prefix: &str) -> proc_macro2::TokenStream 
 }
 
 pub fn convert_idl_type_to_syn_type(ty: &IdlType) -> syn::Type {
-    syn::parse_str(&convert_idl_type_to_str(ty)).unwrap()
+    syn::parse_str(&convert_idl_type_to_str(ty, false)).unwrap()
 }
 
 // TODO: Impl `ToString` for `IdlType`
-pub fn convert_idl_type_to_str(ty: &IdlType) -> String {
+pub fn convert_idl_type_to_str(ty: &IdlType, is_const: bool) -> String {
     match ty {
         IdlType::Bool => "bool".into(),
         IdlType::U8 => "u8".into(),
@@ -60,14 +61,14 @@ pub fn convert_idl_type_to_str(ty: &IdlType) -> String {
         IdlType::I128 => "i128".into(),
         IdlType::U256 => "u256".into(),
         IdlType::I256 => "i256".into(),
-        IdlType::Bytes => "Vec<u8>".into(),
-        IdlType::String => "String".into(),
+        IdlType::Bytes => if is_const { "&[u8]" } else { "Vec<u8>" }.into(),
+        IdlType::String => if is_const { "&str" } else { "String" }.into(),
         IdlType::Pubkey => "Pubkey".into(),
-        IdlType::Option(ty) => format!("Option<{}>", convert_idl_type_to_str(ty)),
-        IdlType::Vec(ty) => format!("Vec<{}>", convert_idl_type_to_str(ty)),
+        IdlType::Option(ty) => format!("Option<{}>", convert_idl_type_to_str(ty, is_const)),
+        IdlType::Vec(ty) => format!("Vec<{}>", convert_idl_type_to_str(ty, is_const)),
         IdlType::Array(ty, len) => format!(
             "[{}; {}]",
-            convert_idl_type_to_str(ty),
+            convert_idl_type_to_str(ty, is_const),
             match len {
                 IdlArrayLen::Generic(len) => len.into(),
                 IdlArrayLen::Value(len) => len.to_string(),
@@ -76,7 +77,7 @@ pub fn convert_idl_type_to_str(ty: &IdlType) -> String {
         IdlType::Defined { name, generics } => generics
             .iter()
             .map(|generic| match generic {
-                IdlGenericArg::Type { ty } => convert_idl_type_to_str(ty),
+                IdlGenericArg::Type { ty } => convert_idl_type_to_str(ty, is_const),
                 IdlGenericArg::Const { value } => value.into(),
             })
             .reduce(|mut acc, cur| {
@@ -368,4 +369,72 @@ fn handle_defined_fields<R>(
         },
         _ => unit_cb(),
     }
+}
+
+/// Combine regular instruction accounts with non-instruction composite accounts.
+pub fn get_all_instruction_accounts(idl: &Idl) -> Vec<IdlInstructionAccounts> {
+    // It's possible to declare an accounts struct and not use it as an instruction, see
+    // https://github.com/coral-xyz/anchor/issues/3274
+    //
+    // NOTE: Returned accounts will not be unique if non-instruction composite accounts have been
+    // used multiple times https://github.com/solana-foundation/anchor/issues/3349
+    fn get_non_instruction_composite_accounts<'a>(
+        accs: &'a [IdlInstructionAccountItem],
+        idl: &'a Idl,
+    ) -> Vec<&'a IdlInstructionAccounts> {
+        accs.iter()
+            .flat_map(|acc| match acc {
+                IdlInstructionAccountItem::Composite(accs)
+                    if !idl
+                        .instructions
+                        .iter()
+                        .any(|ix| ix.accounts == accs.accounts) =>
+                {
+                    let mut nica = get_non_instruction_composite_accounts(&accs.accounts, idl);
+                    nica.push(accs);
+                    nica
+                }
+                _ => Default::default(),
+            })
+            .collect()
+    }
+
+    let ix_accs = idl
+        .instructions
+        .iter()
+        .flat_map(|ix| ix.accounts.to_owned())
+        .collect::<Vec<_>>();
+    get_non_instruction_composite_accounts(&ix_accs, idl)
+        .into_iter()
+        .fold(Vec::<IdlInstructionAccounts>::default(), |mut all, accs| {
+            // Make sure they are unique
+            if all.iter().all(|a| a.accounts != accs.accounts) {
+                // The name is not guaranteed to be the same as the one used in the actual source
+                // code of the program because the IDL only stores the field names
+                let name = if all.iter().all(|a| a.name != accs.name) {
+                    accs.name.to_owned()
+                } else {
+                    // Append numbers to the field name until we find a unique name
+                    (2..)
+                        .find_map(|i| {
+                            let name = format!("{}{i}", accs.name);
+                            all.iter().all(|a| a.name != name).then_some(name)
+                        })
+                        .expect("Should always find a valid name")
+                };
+
+                all.push(IdlInstructionAccounts {
+                    name,
+                    accounts: accs.accounts.to_owned(),
+                })
+            }
+
+            all
+        })
+        .into_iter()
+        .chain(idl.instructions.iter().map(|ix| IdlInstructionAccounts {
+            name: ix.name.to_owned(),
+            accounts: ix.accounts.to_owned(),
+        }))
+        .collect()
 }