+++ lang/syn/src/idl/accounts.rs
@@ -128,7 +128,7 @@ pub fn gen_idl_build_impl_accounts_struct(accounts: &AccountsStruct) -> TokenStr
                     if let Some(ty) = <#defined>::create_type() {
                         let account = #idl::IdlAccount {
                             name: ty.name.clone(),
-                            discriminator: #defined::DISCRIMINATOR.into(),
+                            discriminator: <#defined>::DISCRIMINATOR.into(),
                         };
                         accounts.insert(account.name.clone(), account);
                         types.insert(ty.name.clone(), ty);
@@ -175,7 +175,7 @@ fn get_address(acc: &Field) -> TokenStream {
                         .ident
                         .to_string()
                         .chars()
-                        .all(|c| c.is_uppercase() || c == '_'),
+                        .all(|c| c.is_ascii_uppercase() || c.is_ascii_digit() || c == '_'),
                     // Allow `const fn`s (assume any stand-alone function call without an argument)
                     // e.g. `crate::id()`
                     syn::Expr::Call(expr) => expr.args.is_empty(),